+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- Mocked type definitions from `vtex.product-context` in favor of the actual ones.
 
 ## [1.6.0] - 2020-09-04 [YANKED]
 ### Added